requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.29
google-generativeai==0.4.1
python-dotenv==1.0.1
requests==2.31.0
pandas==2.2.2
openpyxl==3.1.2
Werkzeug==3.0.1